requirements.txt
mysql-connector-python>=8.0.0
faker>=8.0.0
python-dotenv>=0.19.0
networkx>=2.6.0
